pyproject.toml
[tool.poetry]
name = "dungeon-generator"
version = "0.1.0"
description = ""
authors = ["ratilmii <ratilmii@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
numpy = "1.26.4"
pygame-ce = "^2.5.5"
pygame-gui = "^0.6.14"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.2"
coverage = "^7.10.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
